requirements.txt
Django==3.2.25
djangorestframework==3.12.4
django-extensions==3.1.0
psycopg2-binary==2.9.1
django-filter==2.4.0
factory-boy==3.1.0
pytest-django==4.4.0
djangorestframework-simplejwt==5.3.1
black==24.4.2
python-memcached==1.62
freezegun~=1.5.0